chore(publicip): Add schema/types
